requirements.txt
selenium==3.141.0
colorgram-py==1.2.0
pytesseract==0.3.10
pandas==1.5.3